Remove grantlee_add_plugin macro. Add macro to adjust plugin name.
The grantlee_add_plugin macro predates automoc, and was partly
added to simplify moc'ing. Now we expect downstreams to use
CMAKE_AUTOMOC or to otherwise handle the moc issue themselves.

Additionally, make downstream linking to grantlee_core more explicit
by not hiding it in a macro.

The code to add all sources to a global property is specific to
the Grantlee build for coverage analysis, and is not relevant to
downstreams anyway.